feat: form with controlled elements
Created a page at the "/login-controlled" route and a login form with the username and password input and print its data when the user clicks on a button. updated a data object which is updated when the user enters some information inside the form elements. On submission, the data object is filled with the value contained in the form elements.